requirements.txt
# AMD-GAN Python Dependencies
numpy>=1.21.0
pandas>=1.3.0
polars>=0.18.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
lightgbm>=3.3.0
xgboost>=1.5.0
imbalanced-learn>=0.9.0
matplotlib>=3.4.0
scipy>=1.7.0
psutil>=5.8.0
pynvml>=11.0.0
python-dotenv>=0.19.0
